Renamespace decorated responses
